fix: audit sweep — blog filter, admin cleanup, security, UX polish
Blog:
- Extract category filter + posts grid into BlogContent client component
- Category buttons now actually filter by selected category (were decorative)
- Show empty state when no posts match selected category
- Article titles are now links (better UX, same destination as Read arrow)
- Replace text-glyph "->" arrows with lucide ArrowRight (CLAUDE.md design standard)
- Hover border colour now matches category colour

Admin dashboard:
- Remove duplicate Revenue + Moderation cards introduced by merge conflict
- Remove hardcoded $0 MRR stat (was fake — no Stripe data available server-side)
- Fix Moderation card icon from ShieldCheck to ShieldAlert (correct semantics)
- Add href to Moderation card (/admin/moderation)
- Remove fake pendingReviews calculation (Math.floor(total/5))
- Remove unused ShieldCheck import

For Therapists page:
- Replace vague "Growing" stat with honest "Join Early" framing

Signup:
- Add rel="noopener noreferrer" to all target="_blank" legal-doc links

Security:
- Delete src/components/sections/EnhancedBlog.tsx — unused component with
  unsanitized dangerouslySetInnerHTML on user content (XSS vector)
- Remove its re-export from src/components/index.ts
