+++ src/components/index.ts
@@ -13,5 +13,4 @@ export {
   type TherapistProfile as ComparisonTherapistProfile,
   type TherapistComparisonProps,
 } from "@/components/sections/TherapistComparison";
-export { BlogGrid, BlogPostCard, BlogPostDetail, type BlogPost } from "@/components/sections/EnhancedBlog";
 export { PremiumIcon } from "@/components/icons/PremiumIcon";
